🚚 -moved all tsx files into two categories directory : pages, components
